refactor(BaseFragment): Move BaseFragment to coreui. Abstract crashlytics events from base fragment.
