+++ app/src/main/java/piuk/blockchain/android/ui/backup/completed/BackupWalletCompletedFragment.kt
@@ -13,7 +13,7 @@ import piuk.blockchain.android.R
 import piuk.blockchain.android.injection.Injector
 import piuk.blockchain.android.ui.backup.start.BackupWalletStartingFragment
 import piuk.blockchain.android.ui.backup.transfer.ConfirmFundsTransferDialogFragment
-import piuk.blockchain.android.ui.base.BaseFragment
+import piuk.blockchain.androidcoreui.ui.base.BaseFragment
 import piuk.blockchain.androidcoreui.utils.extensions.gone
 import piuk.blockchain.androidcoreui.utils.extensions.inflate
 import java.text.SimpleDateFormat
