+++ app/src/main/java/piuk/blockchain/android/ui/backup/start/BackupWalletStartingFragment.kt
@@ -10,7 +10,7 @@ import piuk.blockchain.android.R
 import piuk.blockchain.android.injection.Injector
 import piuk.blockchain.android.ui.account.SecondPasswordHandler
 import piuk.blockchain.android.ui.backup.wordlist.BackupWalletWordListFragment
-import piuk.blockchain.android.ui.base.BaseFragment
+import piuk.blockchain.androidcoreui.ui.base.BaseFragment
 import piuk.blockchain.androidcoreui.utils.extensions.inflate
 import javax.inject.Inject
