+++ app/src/main/java/piuk/blockchain/android/ui/backup/transfer/ConfirmFundsTransferDialogFragment.kt
@@ -15,7 +15,7 @@ import piuk.blockchain.android.R
 import piuk.blockchain.android.injection.Injector
 import piuk.blockchain.android.ui.account.SecondPasswordHandler
 import piuk.blockchain.android.ui.balance.BalanceFragment
-import piuk.blockchain.android.ui.base.BaseDialogFragment
+import piuk.blockchain.androidcoreui.ui.base.BaseDialogFragment
 import piuk.blockchain.androidcoreui.ui.customviews.MaterialProgressDialog
 import piuk.blockchain.androidcoreui.ui.customviews.ToastCustom
 import piuk.blockchain.androidcoreui.utils.extensions.gone
