+++ app/src/main/java/piuk/blockchain/android/ui/dashboard/DashboardFragment.kt
@@ -17,18 +17,15 @@ import piuk.blockchain.android.data.websocket.WebSocketService
 import piuk.blockchain.android.injection.Injector
 import piuk.blockchain.android.ui.balance.BalanceFragment
 import piuk.blockchain.android.ui.base.BaseAuthActivity
-import piuk.blockchain.android.ui.base.BaseFragment
 import piuk.blockchain.android.ui.charts.ChartsActivity
 import piuk.blockchain.android.ui.customviews.BottomSpacerDecoration
 import piuk.blockchain.android.ui.dashboard.adapter.DashboardDelegateAdapter
 import piuk.blockchain.android.ui.home.MainActivity
-import piuk.blockchain.android.ui.home.MainActivity.ACCOUNT_EDIT
-import piuk.blockchain.android.ui.home.MainActivity.ACTION_RECEIVE_BCH
-import piuk.blockchain.android.ui.home.MainActivity.CONTACTS_EDIT
-import piuk.blockchain.android.ui.home.MainActivity.SETTINGS_EDIT
+import piuk.blockchain.android.ui.home.MainActivity.*
 import piuk.blockchain.android.util.OSUtil
 import piuk.blockchain.androidcore.data.currency.CryptoCurrencies
 import piuk.blockchain.androidcore.utils.helperfunctions.unsafeLazy
+import piuk.blockchain.androidcoreui.ui.base.BaseFragment
 import piuk.blockchain.androidcoreui.utils.AndroidUtils
 import piuk.blockchain.androidcoreui.utils.ViewUtils
 import piuk.blockchain.androidcoreui.utils.extensions.inflate
