+++ coreui/src/main/java/piuk/blockchain/androidcoreui/ui/base/BaseFragment.java
@@ -1,31 +1,19 @@
-package piuk.blockchain.android.ui.base;
+package piuk.blockchain.androidcoreui.ui.base;
 
-import android.os.Bundle;
 import android.support.annotation.CallSuper;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.support.v4.view.ViewPager;
 
 import com.crashlytics.android.answers.ContentViewEvent;
 
 import piuk.blockchain.androidcoreui.utils.logging.Logging;
-import piuk.blockchain.androidcoreui.ui.base.BasePresenter;
-import piuk.blockchain.androidcoreui.ui.base.View;
 
+/**
+ * Logs Fragments that have been visited for statistics purposes using Crashlytics' answers.
+ */
 public abstract class BaseFragment<VIEW extends View, PRESENTER extends BasePresenter<VIEW>>
-        extends Fragment {
+        extends BaseMvpFragment<VIEW, PRESENTER> {
 
-    private PRESENTER presenter;
     private boolean logged;
-
-    @CallSuper
-    @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        presenter = createPresenter();
-        presenter.initView(getMvpView());
-    }
 
     @CallSuper
     @Override
@@ -62,31 +50,5 @@ public abstract class BaseFragment<VIEW extends View, PRESENTER extends BasePres
             }
         }
     }
-
-    @CallSuper
-    @Override
-    public void onPause() {
-        super.onPause();
-        presenter.onViewPaused();
-    }
-
-    @CallSuper
-    @Override
-    public void onDestroy() {
-        super.onDestroy();
-        presenter.onViewDestroyed();
-    }
-
-    protected void onViewReady() {
-        presenter.onViewReady();
-    }
-
-    protected PRESENTER getPresenter() {
-        return presenter;
-    }
-
-    protected abstract PRESENTER createPresenter();
-
-    protected abstract VIEW getMvpView();
-
 }
+
